Make limit a size_t in JasonAsm.

+++ JasonAsm.h
@@ -1,13 +1,13 @@
 #include <cstdint>
 #include <cstring>
 
-static inline int JSONStringCopyInline (uint8_t*& dst, uint8_t const*& src,
-                                        int limit) {
+static inline size_t JSONStringCopyInline (uint8_t*& dst, uint8_t const*& src,
+                                           size_t limit) {
   // Copy up to limit uint8_t from src to dst.
   // Stop at the first control character or backslash or double quote.
   // Report the number of bytes copied. May copy less bytes, for example
   // for alignment reasons.
-  int count = limit;
+  size_t count = limit;
   while (count > 0 && 
          *src >= 32 && 
          *src != '\\' && 
@@ -18,20 +18,20 @@ static inline int JSONStringCopyInline (uint8_t*& dst, uint8_t const*& src,
   return limit - count;
 }
 
-int JSONStringCopyC (uint8_t*& dst, uint8_t const*& src, int limit);
-extern int (*JSONStringCopy)(uint8_t*&, uint8_t const*&, int);
+size_t JSONStringCopyC (uint8_t*& dst, uint8_t const*& src, size_t limit);
+extern size_t (*JSONStringCopy)(uint8_t*&, uint8_t const*&, size_t);
 
 // Now a version which also stops at high bit set bytes:
 
-static inline int JSONStringCopyCheckUtf8Inline (uint8_t*& dst,
-                                                 uint8_t const*& src,
-                                                 int limit) {
+static inline size_t JSONStringCopyCheckUtf8Inline (uint8_t*& dst,
+                                                    uint8_t const*& src,
+                                                    size_t limit) {
   // Copy up to limit uint8_t from src to dst.
   // Stop at the first control character or backslash or double quote.
   // Also stop at byte with high bit set.
   // Report the number of bytes copied. May copy less bytes, for example
   // for alignment reasons.
-  int count = limit;
+  size_t count = limit;
   while (count > 0 && 
          *src >= 32 && 
          *src != '\\' && 
@@ -43,15 +43,17 @@ static inline int JSONStringCopyCheckUtf8Inline (uint8_t*& dst,
   return limit - count;
 }
 
-int JSONStringCopyCheckUtf8C (uint8_t*& dst, uint8_t const*& src, int limit);
-extern int (*JSONStringCopyCheckUtf8)(uint8_t*&, uint8_t const*&, int);
+size_t JSONStringCopyCheckUtf8C (uint8_t*& dst, uint8_t const*& src,
+                                 size_t limit);
+extern size_t (*JSONStringCopyCheckUtf8)(uint8_t*&, uint8_t const*&, size_t);
 
 // White space skipping:
 
-static inline int JSONSkipWhiteSpaceInline (uint8_t const*& ptr, int limit) {
+static inline size_t JSONSkipWhiteSpaceInline (uint8_t const*& ptr,
+                                               size_t limit) {
   // Skip up to limit uint8_t from ptr as long as they are whitespace.
   // Advance ptr and return the number of skipped bytes.
-  int count = limit;
+  size_t count = limit;
   while (count > 0 &&
          (*ptr == ' ' ||
           *ptr == '\t' ||
@@ -63,6 +65,6 @@ static inline int JSONSkipWhiteSpaceInline (uint8_t const*& ptr, int limit) {
   return limit - count;
 }
 
-int JSONSkipWhiteSpaceC (uint8_t const*& ptr, int limit);
-extern int (*JSONSkipWhiteSpace)(uint8_t const*&, int);
+size_t JSONSkipWhiteSpaceC (uint8_t const*& ptr, size_t limit);
+extern size_t (*JSONSkipWhiteSpace)(uint8_t const*&, size_t);
